+++ RadixSortMsdBenchmark.cpp
@@ -40,10 +40,10 @@ int RadixSortMsdBenchmark(vector<unsigned>& uints)
 	// time how long it takes to sort them:
 	for (int i = 0; i < iterationCount; ++i)
 	{
-		for (unsigned int j = 0; j < uints.size(); j++) {	// copy the original random array into the source array each time, since ParallelMergeSort modifies the source array while sorting
+		for (size_t j = 0; j < uints.size(); j++) {	// copy the original random array into the source array each time, since ParallelMergeSort modifies the source array while sorting
 			//uints[j] = j + 2;							// for pre-sorted array testing
 			uintsCopy[j] = uints[j];
-			//sorted[j] = j;									// page in the destination array into system memory
+			//sorted[j] = (unsigned)j;									// page in the destination array into system memory
 		}
 		// Eliminate compiler ability to optimize paging-in of the input and output arrays
 		// Paging-in source and destination arrays leads to a 50% speed-up on Linux, and 15% on Windows
